# openapi/koru_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "koru",
    "description": "An API for managing long-term shared expenses.",
    "contact": {
      "name": "Raphael Biland",
      "email": "raphael.biland@outlook.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.4.0"
  },
  "paths": {
    "/admin/groups": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Fetches all groups, only available if user is ADMIN",
        "description": "Fetches all groups, only available if user is ADMIN\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExamples:\n```\ncurl -i -b cookie \"http://localhost:8000/admin/groups\"\n```\n",
        "operationId": "get_all_groups",
        "responses": {
          "200": {
            "description": "List of all groups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllGroupsResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not Admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Fetches all users, only available if user is ADMIN",
        "description": "Fetches all users, only available if user is ADMIN\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExamples:\n```\ncurl -i -b cookie \"http://localhost:8000/admin/users\"\n```\n",
        "operationId": "get_all_users",
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllUsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User is not Admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Fetches the groups of the user making the request.",
        "description": "Fetches the groups of the user making the request.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExamples:\n```\ncurl -i -b cookie \"http://localhost:8000/groups\"\n```\n",
        "operationId": "get_groups",
        "responses": {
          "200": {
            "description": "List of groups of the user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGroupsResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Creates a new group with the member making the request as admin.",
        "description": "Creates a new group with the member making the request as admin.\n\nRequires the auth cookie from `/login` to be attached to the request.\n\nExample:\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"name\":\"my group\",\"color\":{\"red\":0,\"green\":255,\"blue\":0}}' -b cookie \"http://localhost:8000/groups\"\n```\n",
        "operationId": "create_group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Group created successfully. Returns the ID of the group created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "User is not logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs in the user given valid credentials.",
        "description": "Logs in the user given valid credentials.\n\nIt will return a `Set-Cookie` header, that should be used in next authenticated requests\n\nExample:\n```\ncurl -i -H 'Content-Type: application/json' -d '{\"password\":\"123\",\"email\":\"r@r1.com\"}' -c cookie \"http://localhost:8000/login\"\n```",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Auth cookie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation errors in login request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllGroupsResponse": {
        "type": "object",
        "required": [
          "groups"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDto"
            }
          }
        }
      },
      "AllUsersResponse": {
        "type": "object",
        "required": [
          "users"
        ],
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetailedUserDto"
            }
          }
        }
      },
...
      "ColorDto": {
        "type": "object",
        "required": [
          "red",
          "green",
          "blue"
        ],
        "properties": {
          "blue": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "green": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          },
          "red": {
            "type": "integer",
            "format": "int32",
            "minimum": 0.0
          }
        }
      },
...
      "CreateGroupPayload": {
        "type": "object",
        "required": [
          "name",
          "color"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "DetailedUserDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "ErrorData": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ErrorData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GetAllGroupsResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AllGroupsResponse"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "GetAllUsersResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AllUsersResponse"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GetGroupsResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroupsResponse"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "GroupDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "members"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDto"
            }
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "GroupsResponse": {
        "type": "object",
        "required": [
          "groups"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDto"
            }
          }
        }
      },
      "IdData": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "IdResponse": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IdData"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "LoginPayload": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "MemberDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "is_admin",
          "color",
          "joined_at"
        ],
        "properties": {
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_admin": {
            "type": "boolean"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "id"
      }
    }
  },
...
}